requirements.txt
pacmap>=0.7
python-annoy>=1.17.3
pandas>=1.5.0
scikit-learn>=1.2.0
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.13.1
scipy>=1.7.0
pytorch>=2.4.0
carla-recourse
dice-ml==0.9
shap==0.40.0
lime==0.2.0